-Removed TestMoveThree
-removed a legacy script
