--- Assets/Scripts/TestMoveThree.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ba3e5119f56c2094eb5f45d3a59023b4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
